backend-hybrid/redaction-service/requirements.txt
pypdfium2>=4.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
Flask>=3.0.0
flask-cors>=4.0.0
